fix some bugs

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: silim <silim@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/01 17:40:38 by silim             #+#    #+#             */
-/*   Updated: 2020/10/16 18:01:39 by silim            ###   ########.fr       */
+/*   Updated: 2020/11/14 13:19:05 by silim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_atoi(const char *str)
 {
-	long	num;
+	unsigned long long	num;
 	int		sign;
 	int		count;
